Made rendererbase only use vec2, added boundingbox2d

+++ pixel_engine/Engine/Renderer/Renderer.cs
@@ -18,8 +18,8 @@ namespace pixel_renderer
             }
             lock (frame)
             {
-                stride = 4 * (Resolution.x * 24 + 31) / 32;
-                if (frame.Length != stride * Resolution.y) frame = new byte[stride * Resolution.y];
+                stride = 4 * ((int)Resolution.x * 24 + 31) / 32;
+                if (frame.Length != stride * Resolution.y) frame = new byte[stride * (int)Resolution.y];
                 
                 IEnumerable<UIComponent> uiComponents = Runtime.Instance.GetStage().GetAllComponents<UIComponent>();
 
@@ -32,7 +32,7 @@ namespace pixel_renderer
         public override void Render(System.Windows.Controls.Image output)
         {
             output.Source = BitmapSource.Create(
-                Resolution.x, Resolution.y, 96, 96, System.Windows.Media.PixelFormats.Bgr24, null,
+                (int)Resolution.x, (int)Resolution.y, 96, 96, System.Windows.Media.PixelFormats.Bgr24, null,
                 frame, stride);
         }
     }
